requirements.txt
openai
pydantic
python-dotenv
pandas
numpy
streamlit
altair
